requirements.txt
Django==1.7
django-mathfilters
django-filter
django-dotenv
django-configurations
django-secure
django-extensions
django-compressor
django-bower
djangorestframework
django-debug-toolbar
markdown
requests
requests_oauthlib
MySQL-python
tvdb-api
python-dateutil
Fabric